# backend/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sishi-youxu",
    "description": "四时有序 (sishi-youxu) — Phase 0~3 完成，67 条 API 全部实现。\n\n支持客户端：Web (Vue 3) / Capacitor (Android/iOS) / 微信小程序。\n\nWebSocket 鉴权：先用 access_token 调 `POST /api/v1/auth/ws-ticket`\n换取一次性 ticket，再用 `wss://host/ws/notifications?token=<ticket>` 连接。\n\n调试提示：点击右上角 `Authorize` 按钮，输入 `Bearer <access_token>` 即可全局鉴权。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/tokens": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "账号登录",
        "description": "支持 password / phone_sms / email_code 三种方式",
        "operationId": "login_api_v1_auth_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login Api V1 Auth Tokens Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/tokens/refresh": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "刷新 Token",
        "description": "用 refresh_token 换取新的 access_token",
        "operationId": "refresh_token_api_v1_auth_tokens_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Refresh Token Api V1 Auth Tokens Refresh Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/wechat/login": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "微信登录",
        "description": "小程序 wx.login() code 换取 JWT",
        "operationId": "wechat_login_api_v1_auth_wechat_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Wechat Login Api V1 Auth Wechat Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/ws-ticket": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "获取 WebSocket Ticket",
        "description": "access_token 换一次性 ws ticket",
        "operationId": "issue_ws_ticket_api_v1_auth_ws_ticket_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Issue Ws Ticket Api V1 Auth Ws Ticket Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/auth/captcha/verify": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "校验图形验证码",
        "description": "校验用户输入的图形验证码是否正确",
        "operationId": "verify_captcha_api_v1_auth_captcha_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptchaVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Verify Captcha Api V1 Auth Captcha Verify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/sms/send": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "发送短信验证码",
        "description": "向指定手机号发送短信验证码",
        "operationId": "send_sms_api_v1_auth_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsCodeSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Send Sms Api V1 Auth Sms Send Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/sms/login": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "短信登录",
        "description": "手机号 + 短信验证码登录",
        "operationId": "login_by_sms_api_v1_auth_sms_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsCodeLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login By Sms Api V1 Auth Sms Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/email/send": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "发送邮箱验证码",
        "description": "向指定邮箱发送验证码邮件",
        "operationId": "send_email_code_api_v1_auth_email_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Send Email Code Api V1 Auth Email Send Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/auth/email/login": {
      "post": {
        "tags": [
          "user-auth"
        ],
        "summary": "邮箱验证码登录",
        "description": "邮箱 + 验证码登录",
        "operationId": "login_by_email_api_v1_auth_email_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login By Email Api V1 Auth Email Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "password",
              "phone_sms",
              "email_code",
              "wechat"
            ],
            "title": "Provider",
            "description": "登录方式"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "认证凭据，结构因 provider 而异"
          }
        },
        "type": "object",
        "required": [
          "provider"
        ],
        "title": "LoginRequest",
        "description": "账号登录请求。\n\nprovider 可选：\n- `password`：账号密码登录（payload.identifier + payload.password）\n- `phone_sms`：手机号+短信验证码（payload.phone + payload.code）\n- `email_code`：邮箱+邮箱验证码（payload.email + payload.code）"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "刷新令牌"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "刷新 Token 请求。"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在 Swagger UI 右上角点击 Authorize，输入 `Bearer <access_token>`。access_token 通过 `POST /api/v1/auth/tokens` 获取。"
      },
      "adminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "管理员 token，通过 `POST /api/v1/admin/auth/tokens` 获取。"
      }
    }
  }
}